+++ ui/app/src/components/GroupsManagement/GroupsManagement.tsx
@@ -103,7 +103,7 @@ export default function GroupsManagement() {
   return (
     <section>
       <GlobalAppToolbar
-        title={<FormattedMessage id="globalMenu.groups" defaultMessage="Groups" />}
+        title={<FormattedMessage id="words.groups" defaultMessage="Groups" />}
         leftContent={
           <Button startIcon={<AddIcon />} variant="outlined" color="primary" onClick={() => setOpenGroupDialog(true)}>
             <FormattedMessage id="sites.createGroup" defaultMessage="Create Group" />
