+++ ui/app/src/components/LogConsoleDetailsDialog/LogConsoleDetailsDialog.tsx
@@ -16,16 +16,10 @@
 
 import Dialog from '@material-ui/core/Dialog';
 import React from 'react';
-import { FormattedMessage } from 'react-intl';
 import { LogEvent } from '../../models/monitoring/LogEvent';
-import { useMount, useSelection } from '../../utils/hooks';
-import DialogHeader from '../Dialogs/DialogHeader';
-import DialogBody from '../Dialogs/DialogBody';
-import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
-import { useStyles } from './styles';
+import LogConsoleDetailsDialogUI from './LogConsoleDetailsDialogUI';
 
-interface LogConsoleDetailsDialogProps {
+export interface LogConsoleDetailsDialogProps {
   open: boolean;
   onClose(): void;
   onClosed?(): void;
@@ -40,74 +34,3 @@ export default function LogConsoleDetailsDialog(props: LogConsoleDetailsDialogPr
     </Dialog>
   );
 }
-
-function LogConsoleDetailsDialogUI(props: LogConsoleDetailsDialogProps) {
-  const { logEvent, onClose, onClosed } = props;
-  const localeBranch = useSelection((state) => state.uiConfig.locale);
-  const classes = useStyles();
-
-  useMount(() => {
-    return () => {
-      onClosed?.();
-    };
-  });
-
-  return (
-    <>
-      <DialogHeader
-        title={<FormattedMessage id="logConsoleDetailsDialog.title" defaultMessage="Log Details" />}
-        onDismiss={onClose}
-      />
-      <DialogBody>
-        <Box display="flex" p="10px 0" alignItems="center">
-          <Typography color="textSecondary" className={classes.label}>
-            <FormattedMessage id="words.level" defaultMessage="Level" />
-          </Typography>
-          <Box display="flex" width="100%" alignItems="center">
-            <Typography>{logEvent.level}</Typography>
-          </Box>
-        </Box>
-        <Box display="flex" p="10px 0" alignItems="center">
-          <Typography color="textSecondary" className={classes.label}>
-            <FormattedMessage id="words.timestamp" defaultMessage="timestamp" />
-          </Typography>
-          <Box display="flex" width="100%" alignItems="center">
-            <Typography>
-              {new Intl.DateTimeFormat(localeBranch.localeCode, localeBranch.dateTimeFormatOptions).format(
-                new Date(logEvent.timestamp)
-              )}
-            </Typography>
-          </Box>
-        </Box>
-        <Box display="flex" p="10px 0" alignItems="center">
-          <Typography color="textSecondary" className={classes.label}>
-            <FormattedMessage id="words.thread" defaultMessage="Thread" />
-          </Typography>
-          <Box display="flex" width="100%" alignItems="center">
-            <Typography>{logEvent.thread}</Typography>
-          </Box>
-        </Box>
-        <Box display="flex" p="10px 0" alignItems="center">
-          <Typography color="textSecondary" className={classes.label}>
-            <FormattedMessage id="words.site" defaultMessage="Site" />
-          </Typography>
-          <Box display="flex" width="100%" alignItems="center">
-            <Typography>{logEvent.site}</Typography>
-          </Box>
-        </Box>
-        <Box display="flex" p="10px 0" alignItems="center">
-          <Typography color="textSecondary" className={classes.label}>
-            <FormattedMessage id="words.message" defaultMessage="Message" />
-          </Typography>
-          <Box display="flex" width="100%" alignItems="center">
-            <Typography>{logEvent.message}</Typography>
-          </Box>
-        </Box>
-      </DialogBody>
-    </>
-  );
-}
-//
-// width: '100%',
-//   display: 'flex',
-//   alignItems: 'center'
