+++ ui/app/src/components/UsersManagement/UsersManagement.tsx
@@ -106,7 +106,7 @@ export default function UsersManagement(props: UsersManagementProps) {
   return (
     <section>
       <GlobalAppToolbar
-        title={<FormattedMessage id="globalMenu.users" defaultMessage="Users" />}
+        title={<FormattedMessage id="words.users" defaultMessage="Users" />}
         leftContent={
           <Button
             startIcon={<AddIcon />}
@@ -114,7 +114,7 @@ export default function UsersManagement(props: UsersManagementProps) {
             color="primary"
             onClick={() => setOpenCreateUserDialog(true)}
           >
-            <FormattedMessage id="sites.createUser" defaultMessage="Create User" />
+            <FormattedMessage id="usersGrid.createUser" defaultMessage="Create User" />
           </Button>
         }
       />
