+++ ui/app/src/translations/locales/de.json
@@ -2,14 +2,14 @@
   "GlobalMenu.AboutUs": "Über",
   "GlobalMenu.AuditEntryLabel": "PRÜFUNG",
   "GlobalMenu.ClusterEntryLabel": "Cluster",
-  "GlobalMenu.EncryptionTool": "Verschlüsselungs-Tool",
-  "GlobalMenu.GlobalConfigEntryLabel": "Globale Konfiguration",
+  "globalMenu.encryptionTool": "Verschlüsselungs-Tool",
+  "globalMenu.globalConfigEntryLabel": "Globale Konfiguration",
   "GlobalMenu.GroupsEntryLabel": "Gruppen",
-  "GlobalMenu.LogConsoleEntryLabel": "Protokollkonsole",
-  "GlobalMenu.LoggingLevelsEntryLabel": "Protokollierungsstufen",
+  "globalMenu.logConsoleEntryLabel": "Protokollkonsole",
+  "globalMenu.loggingLevelsEntryLabel": "Protokollierungsstufen",
   "GlobalMenu.Login": "Anmeldung",
-  "GlobalMenu.Recover": "Passwort-Wiederherstellung",
-  "GlobalMenu.Settings": "Kontoverwaltung",
+  "globalMenu.recover": "Passwort-Wiederherstellung",
+  "globalMenu.settings": "Kontoverwaltung",
   "GlobalMenu.SitesEntryLabel": "Websites",
   "GlobalMenu.UsersEntryLabel": "Benutzer",
   "aboutView.attribution": "Crafter CMS wird durch diese anderen <a>Quelloffene Software projekte</a> ermöglicht.",
