+++ ui/app/src/translations/locales/en.json
@@ -2,14 +2,14 @@
   "GlobalMenu.AboutUs": "About",
   "GlobalMenu.AuditEntryLabel": "Audit",
   "GlobalMenu.ClusterEntryLabel": "Cluster",
-  "GlobalMenu.EncryptionTool": "Encryption Tool",
-  "GlobalMenu.GlobalConfigEntryLabel": "Global Config",
+  "globalMenu.encryptionTool": "Encryption Tool",
+  "globalMenu.globalConfigEntryLabel": "Global Config",
   "GlobalMenu.GroupsEntryLabel": "Groups",
-  "GlobalMenu.LogConsoleEntryLabel": "Log Console",
-  "GlobalMenu.LoggingLevelsEntryLabel": "Logging Levels",
+  "globalMenu.logConsoleEntryLabel": "Log Console",
+  "globalMenu.loggingLevelsEntryLabel": "Logging Levels",
   "GlobalMenu.Login": "Login",
-  "GlobalMenu.Recover": "Password Recovery",
-  "GlobalMenu.Settings": "Account",
+  "globalMenu.recover": "Password Recovery",
+  "globalMenu.settings": "Account",
   "GlobalMenu.SitesEntryLabel": "Sites",
   "GlobalMenu.UsersEntryLabel": "Users",
   "aboutView.attribution": "Crafter CMS is made possible by these other <a>open source software projects</a>.",
