+++ ui/app/src/translations/locales/es.json
@@ -2,14 +2,14 @@
   "GlobalMenu.AboutUs": "Acerca de",
   "GlobalMenu.AuditEntryLabel": "Auditoría",
   "GlobalMenu.ClusterEntryLabel": "Clúster",
-  "GlobalMenu.EncryptionTool": "Herramienta de cifrado",
-  "GlobalMenu.GlobalConfigEntryLabel": "Configuración Global",
+  "globalMenu.encryptionTool": "Herramienta de cifrado",
+  "globalMenu.globalConfigEntryLabel": "Configuración Global",
   "GlobalMenu.GroupsEntryLabel": "Grupos",
-  "GlobalMenu.LogConsoleEntryLabel": "Consola de registro",
-  "GlobalMenu.LoggingLevelsEntryLabel": "Niveles de registro",
+  "globalMenu.logConsoleEntryLabel": "Consola de registro",
+  "globalMenu.loggingLevelsEntryLabel": "Niveles de registro",
   "GlobalMenu.Login": "Iniciar sesión",
-  "GlobalMenu.Recover": "Recuperación de contraseña",
-  "GlobalMenu.Settings": "Administración de cuentas",
+  "globalMenu.recover": "Recuperación de contraseña",
+  "globalMenu.settings": "Administración de cuentas",
   "GlobalMenu.SitesEntryLabel": "Sitios",
   "GlobalMenu.UsersEntryLabel": "Usuarios",
   "aboutView.attribution": "Crafter CMS es posible gracias a otros <a>proyectos de software de código abierto</a>.",
