+++ ui/app/src/translations/locales/ko.json
@@ -2,14 +2,14 @@
   "GlobalMenu.AboutUs": "약",
   "GlobalMenu.AuditEntryLabel": "심사",
   "GlobalMenu.ClusterEntryLabel": "클러스터",
-  "GlobalMenu.EncryptionTool": "암호화 도구",
-  "GlobalMenu.GlobalConfigEntryLabel": "글로벌 구성",
+  "globalMenu.encryptionTool": "암호화 도구",
+  "globalMenu.globalConfigEntryLabel": "글로벌 구성",
   "GlobalMenu.GroupsEntryLabel": "여러 떼",
-  "GlobalMenu.LogConsoleEntryLabel": "로그 콘솔",
-  "GlobalMenu.LoggingLevelsEntryLabel": "로깅 수준",
+  "globalMenu.logConsoleEntryLabel": "로그 콘솔",
+  "globalMenu.loggingLevelsEntryLabel": "로깅 수준",
   "GlobalMenu.Login": "로그인",
-  "GlobalMenu.Recover": "비밀번호 복구",
-  "GlobalMenu.Settings": "계정 관리",
+  "globalMenu.recover": "비밀번호 복구",
+  "globalMenu.settings": "계정 관리",
   "GlobalMenu.SitesEntryLabel": "사이트",
   "GlobalMenu.UsersEntryLabel": "사용자",
   "aboutView.attribution": "Crafter CMS는 여러가지 <a>Open Source Software Projects</a>와 연동하여 기능합니다",
